DB 테이블명 변경 (추후 다시 변경할 예정)

+++ backend/routers/policy_router.py
@@ -1,4 +1,5 @@
 from fastapi import APIRouter, HTTPException, Query, Depends
+from sqlalchemy import Integer
 from sqlalchemy.orm import Session, joinedload
 from typing import Optional, List
 from datetime import date
@@ -89,8 +90,8 @@ def optimize_policies(
         db.query(Policy)
         .options(joinedload(Policy.tiers))
         .filter(Policy.is_active == True)
-        .filter((Policy.age_min == None) | (Policy.age_min <= age))
-        .filter((Policy.age_max == None) | (Policy.age_max >= age))
+        .filter((Policy.age_min == None) | (Policy.age_min.cast(Integer) <= age))
+        .filter((Policy.age_max == None) | (Policy.age_max.cast(Integer) >= age))
     )
 
     if request.profile.is_employed:
